+++ src/app/locales/en/components/-storage/swift.js
@@ -1,9 +1,11 @@
 export default {
   username: { name: 'Admin username' },
   password: { name: 'Admin password' },
+  projectName: { name: 'Project name' },
+  userDomainName: { name: 'User domain name' },
+  projectDomainName: { name: 'Project domain name' },
   authUrl: { name: 'Auth URL' },
   containerName: { name: 'Container name' },
-  tenantName: { name: 'Tenant name' },
   blockSize: { name: 'Block size' },
   timeout: { name: 'Timeout [ms]' },
 };